AutomaticWorld/requirements.txt
blessings==1.6
bpython==0.16
certifi==2017.7.27.1
chardet==3.0.4
curtsies==0.2.11
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.34.0
django-rest-auth==0.9.2
djangorestframework==3.6.4
greenlet==0.4.12
gunicorn==19.7.1
heroku==0.1.4
idna==2.6
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
Pygments==2.2.0
python-dateutil==1.5
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
